requirements.txt
absl-py==2.1.0
alembic==1.13.1
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
dnspython==2.6.1
executing==2.0.1
filelock==3.13.4
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
fonttools==4.51.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grpcio==1.62.1
hydra-core==1.3.2
idna==3.7
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
kiwisolver==1.4.5
Mako==1.3.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.3
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pillow==10.3.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.29
stack-data==0.6.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.0.3
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
